+++ test/common/functions.py
@@ -1,0 +1,81 @@
+
+def data_layout(val: int) -> LayoutLike:
+    return [("data", val)]
+
+
+def set_inputs(values: RecordValueDict, field: Record) -> TestGen[None]:
+    for name, value in values.items():
+        if isinstance(value, dict):
+            yield from set_inputs(value, getattr(field, name))
+        else:
+            yield getattr(field, name).eq(value)
+
+
+def get_outputs(field: Record) -> TestGen[RecordIntDict]:
+    # return dict of all signal values in a record because amaranth's simulator can't read all
+    # values of a Record in a single yield - it can only read Values (Signals)
+    result = {}
+    for name, _, _ in field.layout:
+        val = getattr(field, name)
+        if isinstance(val, Signal):
+            result[name] = yield val
+        else:  # field is a Record
+            result[name] = yield from get_outputs(val)
+    return result
+
+
+def neg(x: int, xlen: int) -> int:
+    """
+    Computes the negation of a number in the U2 system.
+
+    Parameters
+    ----------
+    x: int
+        Number in U2 system.
+    xlen : int
+        Bit width of x.
+
+    Returns
+    -------
+    return : int
+        Negation of x in the U2 system.
+    """
+    return (-x) & (2**xlen - 1)
+
+
+def int_to_signed(x: int, xlen: int) -> int:
+    """
+    Converts a Python integer into its U2 representation.
+
+    Parameters
+    ----------
+    x: int
+        Signed Python integer.
+    xlen : int
+        Bit width of x.
+
+    Returns
+    -------
+    return : int
+        Representation of x in the U2 system.
+    """
+    return x & (2**xlen - 1)
+
+
+def signed_to_int(x: int, xlen: int) -> int:
+    """
+    Changes U2 representation into Python integer
+
+    Parameters
+    ----------
+    x: int
+        Number in U2 system.
+    xlen : int
+        Bit width of x.
+
+    Returns
+    -------
+    return : int
+        Representation of x as signed Python integer.
+    """
+    return x | -(x & (2 ** (xlen - 1)))
